+++ NEWS.md
@@ -1,5 +1,7 @@
 ## Version 1.0.3 Release Notes
 - Fixed bugs for CRAN checking
+- Update kegg.RData (November, 2021)
+- Added kegg.pathways.RData (November, 2021)
 - Added pkgdown website
 
 ## Version 1.0.2 Release Notes